Add additional attributes to divider
[MAILPOET-2835]
